Implement query parameter routing for edit preprints.
