+++ app/router.js
@@ -23,10 +23,10 @@ const Router = Ember.Router.extend({
 
 Router.map(function() {
     this.route('index', {path: 'preprints'});
-    this.route('submit', { path: '/:preprint_id/edit'});
-    this.route('submit', {path: 'preprints/submit'}); // Add route needs to be second so it is default.
+    this.route('submit', {path: 'preprints/submit'});
     this.route('discover', {path: 'preprints/discover'});
     this.route('content', { path: '/:preprint_id' });
+    this.route('page-not-found', { path: '/:preprint_id/*bad_url'});
     this.route('page-not-found', { path: 'preprints/*bad_url'});
     this.route('page-not-found', {path: 'preprints/page-not-found'});
     this.route('forbidden');
