+++ app/templates/components/preprint-navbar.hbs
@@ -9,7 +9,7 @@
         </div>
         <div id="preprint-navbar-links" class="navbar-collapse collapse navbar-right">
             <ul class="nav navbar-nav">
-                <li><a href="/preprints/submit/" class="" invokeAction={{action "click" "link" "Navbar - Add preprint"}}>{{t "global.add_preprint"}}</a></li>
+                <li>{{#link-to "submit" class="" invokeAction=(action "click" "link" "Navbar - Add preprint")}}{{t "global.add_preprint"}}{{/link-to}}</li>
                 <li>{{#link-to "discover" class="" invokeAction=(action "click" "link" "Navbar - Search")}}{{t "global.search"}}{{/link-to}}</li>
             </ul>
         </div>
